+++ pertpy/__init__.py
@@ -11,4 +11,4 @@ pertpy_pypi_latest = PypiLatest("pertpy", __version__)
 pertpy_pypi_latest.check_latest()
 
 from . import data as dt
-from . import tools as tl
+from . import tl as tl
