--- pertpy/tools/__init__.py
@@ -1,1 +1,0 @@
-from pertpy.tools._example import add_me
